move util inside lib
